requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Vector DB
lancedb>=0.4.0

# PDF Processing (robust fallback chain)
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# DOCX Processing
python-docx>=1.1.0

# Image Processing
Pillow>=10.0.0
easyocr>=1.7.0

# Audio/Video
openai-whisper>=20231117
moviepy>=1.0.3
opencv-python>=4.8.0
yt-dlp>=2024.1.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=2.1.0

# Utilities
numpy>=1.24.0
loguru>=0.7.0
